retain go1.3 backwards compatibility
